runpod/requirements.txt
# Core dependencies
# torch, torchvision, torchaudio installed in Dockerfile
numpy<2
opencv-python
imageio
imageio-ffmpeg
av
# decord - skip for now, not available for aarch64
einops
tensorboard
tqdm
omegaconf
transformers
diffusers
accelerate
xformers
ftfy
gradio
soundfile
moviepy
oss2

# SeedVR specific dependencies
# flash_attn installed separately in Dockerfile
# apex installed separately in Dockerfile

# RunPod
runpod

# Additional utilities
boto3  # For S3 uploads
requests
Pillow
scipy

# Google Cloud Storage
google-cloud-storage==2.14.0
google-auth==2.27.0